Fix diff, move test to beginning of read_table
