+++ ffi/examples/read-table/read_table.c
@@ -201,7 +201,7 @@ PartitionList* get_partition_list(SharedGlobalScanState* state)
 int main(int argc, char* argv[])
 {
   test_kernel_expr();
-  return -1;
+
   if (argc < 2) {
     printf("Usage: %s table/path\n", argv[0]);
     return -1;
